requirements.txt
flask
anthropic
python-dotenv
pdfplumber
